backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
channels==4.1.0
channels-redis==4.2.0
asgiref==3.8.1
pytest==8.2.0
pytest-django==4.9.0
